add home work 4 with menu

+++ src/OOP/Zoo/factory/LionsFactory.java
@@ -11,7 +11,10 @@ public class LionsFactory {
         ArrayList<Lion> lions = new ArrayList<>();
 
         for (int i = 0; i < lionCount; i++) {
-            lions.add(new Lion(String.format("Lion%s", i + 1), 2000 + i, 4, (int) (10 * Math.random()),
+            lions.add(new Lion(String.format("Lion%s", i + 1),
+                    2000 + i,
+                    4,
+                    (10 * Math.random()),
                     (int) (10 * Math.random())));
         }
