+++ src/OOP/Zoo/factory/WolfsFactory.java
@@ -11,7 +11,10 @@ public class WolfsFactory {
         ArrayList<Wolf> wolfs = new ArrayList<>();
 
         for (int i = 0; i < lionCount; i++) {
-            wolfs.add(new Wolf(String.format("Wolf%s", i + 1), 2000 + i, 4, (int) (10 * Math.random())));
+            wolfs.add(new Wolf(String.format("Wolf%s", i + 1),
+                    2000 + i,
+                    4,
+                    (10 * Math.random())));
         }
 
         return wolfs;
